refactor: minor refactoring in percy.js/proxy.js + some comment fix
